mock_server/requirements.txt
flask
prometheus_client
pyyaml
numpy
